[AIRFLOW-5773] Migrate AWS Athena components to /providers/aws [AIP-21]
