+++ airflow/contrib/sensors/aws_athena_sensor.py
@@ -17,63 +17,15 @@
 # specific language governing permissions and limitations
 # under the License.
 
+"""This module is deprecated. Please use `airflow.providers.aws.sensors.athena`."""
 
-from airflow.contrib.hooks.aws_athena_hook import AWSAthenaHook
-from airflow.exceptions import AirflowException
-from airflow.sensors.base_sensor_operator import BaseSensorOperator
-from airflow.utils.decorators import apply_defaults
+import warnings
 
+# pylint: disable=unused-import
+from airflow.providers.aws.sensors.athena import AthenaSensor  # noqa
 
-class AthenaSensor(BaseSensorOperator):
-    """
-    Asks for the state of the Query until it reaches a failure state or success state.
-    If it fails, failing the task.
-
-    :param query_execution_id: query_execution_id to check the state of
-    :type query_execution_id: str
-    :param max_retires: Number of times to poll for query state before
-        returning the current state, defaults to None
-    :type max_retires: int
-    :param aws_conn_id: aws connection to use, defaults to 'aws_default'
-    :type aws_conn_id: str
-    :param sleep_time: Time to wait between two consecutive call to
-        check query status on athena, defaults to 10
-    :type sleep_time: int
-    """
-
-    INTERMEDIATE_STATES = ('QUEUED', 'RUNNING',)
-    FAILURE_STATES = ('FAILED', 'CANCELLED',)
-    SUCCESS_STATES = ('SUCCEEDED',)
-
-    template_fields = ['query_execution_id']
-    template_ext = ()
-    ui_color = '#66c3ff'
-
-    @apply_defaults
-    def __init__(self,
-                 query_execution_id,
-                 max_retires=None,
-                 aws_conn_id='aws_default',
-                 sleep_time=10,
-                 *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.aws_conn_id = aws_conn_id
-        self.query_execution_id = query_execution_id
-        self.hook = None
-        self.sleep_time = sleep_time
-        self.max_retires = max_retires
-
-    def poke(self, context):
-        self.hook = self.get_hook()
-        self.hook.get_conn()
-        state = self.hook.poll_query_status(self.query_execution_id, self.max_retires)
-
-        if state in self.FAILURE_STATES:
-            raise AirflowException('Athena sensor failed')
-
-        if state in self.INTERMEDIATE_STATES:
-            return False
-        return True
-
-    def get_hook(self):
-        return AWSAthenaHook(self.aws_conn_id, self.sleep_time)
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.aws.sensors.athena`.",
+    DeprecationWarning,
+    stacklevel=2,
+)
